Highlight Plotly Points

This JavaScript code defines a function that is used in a Plotly chart to capture changes in the visibility of the traces on the chart. The layer_brush is used to extract a DOM element from a D3.js selection to access its properties and modify its attributes.

highlight_clicks(plotly_obj, layer_brush = 0)

Arguments

plotly_obj

plotly object with points.

layer_brush

Provide a numeric value. If none supplied, layer_brush = 0 will be used. In shiny, if this value is larger than 3, consider using modules.

Examples

## Not run:
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(ggplot2)

p_1 <- mtcars %>%
  ggplot(aes(x = mpg, y = disp)) +
  geom_point()

library(plotly)
#> 
#> Attaching package: ‘plotly’
#> The following object is masked from ‘package:ggplot2’:
#> 
#>     last_plot
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
#> The following object is masked from ‘package:graphics’:
#> 
#>     layout
library(htmlwidgets)

p_1 %>%
 ggplotly() %>%
 onRender(jsCode = highlight_clicks(.))

## End(Not run)